# Compute webhook selectors and client once