**What type of PR is this?**
/kind cleanup

**What this PR does / why we need it**:
Avoids recomputing the object and namespace selector, and the client config cache key for webhooks on every request.

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

Benchmark setup:
```
go run ./vendor/k8s.io/apiserver/pkg/admission/plugin/webhook/testing/main/main.go
serving on https://127.0.0.1:56558

WEBHOOK_TEST_SERVER_URL=https://127.0.0.1:56558 go test ./vendor/k8s.io/apiserver/pkg/admission/plugin/webhook/mutating -bench . -benchmem -run=NONE > old-mutating.txt 
WEBHOOK_TEST_SERVER_URL=https://127.0.0.1:56558 go test ./vendor/k8s.io/apiserver/pkg/admission/plugin/webhook/validating -bench . -benchmem -run=NONE > old-validating.txt 

WEBHOOK_TEST_SERVER_URL=https://127.0.0.1:56558 go test ./vendor/k8s.io/apiserver/pkg/admission/plugin/webhook/mutating -bench . -benchmem -run=NONE > new-mutating.txt 
WEBHOOK_TEST_SERVER_URL=https://127.0.0.1:56558 go test ./vendor/k8s.io/apiserver/pkg/admission/plugin/webhook/validating -bench . -benchmem -run=NONE > new-validating.txt 
```

results:
```
benchmark                                                                                         old ns/op     new ns/op     delta
BenchmarkValidate/no_match-8                                                                      1075          1111          +3.35%
BenchmarkValidate/match_&_allow-8                                                                 264106        261397        -1.03%
BenchmarkValidate/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns-8        2727          1400          -48.66%
BenchmarkValidate/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns_ii-8     2323          1350          -41.89%
BenchmarkValidate/match_&_allow_(url)-8                                                           260119        263347        +1.24%
BenchmarkValidate/no_match_dry_run-8                                                              1017          1000          -1.67%
BenchmarkValidate/match_dry_run_side_effects_None-8                                               279578        254158        -9.09%
BenchmarkValidate/match_dry_run_side_effects_NoneOnDryRun-8                                       279564        255092        -8.75%
BenchmarkValidate/illegal_annotation_format-8                                                     284790        278731        -2.13%
BenchmarkValidate/skip_webhook_whose_objectSelector_does_not_match-8                              286823        253339        -11.67%
BenchmarkValidate/skip_webhook_whose_objectSelector_does_not_match_CRD's_labels-8                 264151        248464        -5.94%
```

```
benchmark                                                                                         old allocs     new allocs     delta
BenchmarkValidate/no_match-8                                                                      6              6              +0.00%
BenchmarkValidate/match_&_allow-8                                                                 632            624            -1.27%
BenchmarkValidate/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns-8        15             8              -46.67%
BenchmarkValidate/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns_ii-8     15             8              -46.67%
BenchmarkValidate/match_&_allow_(url)-8                                                           631            624            -1.11%
BenchmarkValidate/no_match_dry_run-8                                                              6              6              +0.00%
BenchmarkValidate/match_dry_run_side_effects_None-8                                               632            624            -1.27%
BenchmarkValidate/match_dry_run_side_effects_NoneOnDryRun-8                                       632            624            -1.27%
BenchmarkValidate/illegal_annotation_format-8                                                     659            652            -1.06%
BenchmarkValidate/skip_webhook_whose_objectSelector_does_not_match-8                              642            626            -2.49%
BenchmarkValidate/skip_webhook_whose_objectSelector_does_not_match_CRD's_labels-8                 642            626            -2.49%
```
```
benchmark                                                                                         old bytes     new bytes     delta
BenchmarkValidate/no_match-8                                                                      638           639           +0.16%
BenchmarkValidate/match_&_allow-8                                                                 27341         22390         -18.11%
BenchmarkValidate/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns-8        1091          790           -27.59%
BenchmarkValidate/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns_ii-8     1057          793           -24.98%
BenchmarkValidate/match_&_allow_(url)-8                                                           27228         22305         -18.08%
BenchmarkValidate/no_match_dry_run-8                                                              617           614           -0.49%
BenchmarkValidate/match_dry_run_side_effects_None-8                                               27290         22344         -18.12%
BenchmarkValidate/match_dry_run_side_effects_NoneOnDryRun-8                                       27288         22337         -18.14%
BenchmarkValidate/illegal_annotation_format-8                                                     32086         27208         -15.20%
BenchmarkValidate/skip_webhook_whose_objectSelector_does_not_match-8                              27930         22528         -19.34%
BenchmarkValidate/skip_webhook_whose_objectSelector_does_not_match_CRD's_labels-8                 28094         22680         -19.27%
```

```
benchmark                                                                                      old ns/op     new ns/op     delta
BenchmarkAdmit/match_&_remove_label-8                                                          354432        299584        -15.47%
BenchmarkAdmit/match_&_add_label-8                                                             468388        411575        -12.13%
BenchmarkAdmit/match_CRD_&_add_label-8                                                         301069        270244        -10.24%
BenchmarkAdmit/match_CRD_&_remove_label-8                                                      259746        230004        -11.45%
BenchmarkAdmit/first_webhook_remove_labels,_second_webhook_shouldn't_be_called-8               4993          2662          -46.69%
BenchmarkAdmit/first_webhook_remove_labels_from_CRD,_second_webhook_shouldn't_be_called-8      6015          3472          -42.28%
BenchmarkAdmit/match_&_reinvoke_if_needed_policy-8                                             689181        577692        -16.18%
BenchmarkAdmit/match_&_never_reinvoke_policy-8                                                 428772        399701        -6.78%
BenchmarkAdmit/match_&_never_reinvoke_policy_(by_default)-8                                    443057        401451        -9.39%
BenchmarkAdmit/match_&_no_reinvoke-8                                                           261543        236949        -9.40%
BenchmarkAdmit/no_match-8                                                                      2049          2065          +0.78%
BenchmarkAdmit/match_&_allow-8                                                                 256604        243131        -5.25%
BenchmarkAdmit/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns-8        3550          2517          -29.10%
BenchmarkAdmit/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns_ii-8     3585          2243          -37.43%
BenchmarkAdmit/match_&_allow_(url)-8                                                           271428        248382        -8.49%
BenchmarkAdmit/no_match_dry_run-8                                                              2143          2151          +0.37%
BenchmarkAdmit/match_dry_run_side_effects_None-8                                               277550        298720        +7.63%
BenchmarkAdmit/match_dry_run_side_effects_NoneOnDryRun-8                                       267333        249574        -6.64%
BenchmarkAdmit/illegal_annotation_format-8                                                     274315        263585        -3.91%
BenchmarkAdmit/skip_webhook_whose_objectSelector_does_not_match-8                              252154        261920        +3.87%
BenchmarkAdmit/skip_webhook_whose_objectSelector_does_not_match_CRD's_labels-8                 196856        187236        -4.89%
```
```
benchmark                                                                                      old allocs     new allocs     delta
BenchmarkAdmit/match_&_remove_label-8                                                          786            777            -1.15%
BenchmarkAdmit/match_&_add_label-8                                                             1442           1433           -0.62%
BenchmarkAdmit/match_CRD_&_add_label-8                                                         585            577            -1.37%
BenchmarkAdmit/match_CRD_&_remove_label-8                                                      407            399            -1.97%
BenchmarkAdmit/first_webhook_remove_labels,_second_webhook_shouldn't_be_called-8               28             12             -57.14%
BenchmarkAdmit/first_webhook_remove_labels_from_CRD,_second_webhook_shouldn't_be_called-8      36             20             -44.44%
BenchmarkAdmit/match_&_reinvoke_if_needed_policy-8                                             1343           1327           -1.19%
BenchmarkAdmit/match_&_never_reinvoke_policy-8                                                 1440           1432           -0.56%
BenchmarkAdmit/match_&_never_reinvoke_policy_(by_default)-8                                    1440           1432           -0.56%
BenchmarkAdmit/match_&_no_reinvoke-8                                                           626            618            -1.28%
BenchmarkAdmit/no_match-8                                                                      8              8              +0.00%
BenchmarkAdmit/match_&_allow-8                                                                 637            629            -1.26%
BenchmarkAdmit/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns-8        17             10             -41.18%
BenchmarkAdmit/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns_ii-8     17             10             -41.18%
BenchmarkAdmit/match_&_allow_(url)-8                                                           636            629            -1.10%
BenchmarkAdmit/no_match_dry_run-8                                                              8              8              +0.00%
BenchmarkAdmit/match_dry_run_side_effects_None-8                                               637            629            -1.26%
BenchmarkAdmit/match_dry_run_side_effects_NoneOnDryRun-8                                       637            629            -1.26%
BenchmarkAdmit/illegal_annotation_format-8                                                     664            657            -1.05%
BenchmarkAdmit/skip_webhook_whose_objectSelector_does_not_match-8                              647            631            -2.47%
BenchmarkAdmit/skip_webhook_whose_objectSelector_does_not_match_CRD's_labels-8                 260            244            -6.15%
```
```
benchmark                                                                                      old bytes     new bytes     delta
BenchmarkAdmit/match_&_remove_label-8                                                          40137         33707         -16.02%
BenchmarkAdmit/match_&_add_label-8                                                             56842         50592         -11.00%
BenchmarkAdmit/match_CRD_&_add_label-8                                                         42223         36407         -13.77%
BenchmarkAdmit/match_CRD_&_remove_label-8                                                      32912         27490         -16.47%
BenchmarkAdmit/first_webhook_remove_labels,_second_webhook_shouldn't_be_called-8               2916          2270          -22.15%
BenchmarkAdmit/first_webhook_remove_labels_from_CRD,_second_webhook_shouldn't_be_called-8      3232          2599          -19.59%
BenchmarkAdmit/match_&_reinvoke_if_needed_policy-8                                             74961         64850         -13.49%
BenchmarkAdmit/match_&_never_reinvoke_policy-8                                                 52454         47398         -9.64%
BenchmarkAdmit/match_&_never_reinvoke_policy_(by_default)-8                                    52492         47432         -9.64%
BenchmarkAdmit/match_&_no_reinvoke-8                                                           28033         23018         -17.89%
BenchmarkAdmit/no_match-8                                                                      1919          1915          -0.21%
BenchmarkAdmit/match_&_allow-8                                                                 28487         23529         -17.40%
BenchmarkAdmit/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns-8        2414          2083          -13.71%
BenchmarkAdmit/match_&_disallow_&_but_allowed_because_namespaceSelector_exempt_the_ns_ii-8     2400          2108          -12.17%
BenchmarkAdmit/match_&_allow_(url)-8                                                           28355         23470         -17.23%
BenchmarkAdmit/no_match_dry_run-8                                                              1917          1911          -0.31%
BenchmarkAdmit/match_dry_run_side_effects_None-8                                               28462         23527         -17.34%
BenchmarkAdmit/match_dry_run_side_effects_NoneOnDryRun-8                                       28473         23504         -17.45%
BenchmarkAdmit/illegal_annotation_format-8                                                     33254         28370         -14.69%
BenchmarkAdmit/skip_webhook_whose_objectSelector_does_not_match-8                              28961         23695         -18.18%
BenchmarkAdmit/skip_webhook_whose_objectSelector_does_not_match_CRD's_labels-8                 24262         18931         -21.97%
```

/sig api-machinery
/cc @jpbetz @sttts @roycaihw